# discovery/kubernetes: optimize resolvePodRef