@GiedriusS 2024-10-28
Thanks! I didn't see that function and it makes sense. I have updated other similar code paths to use namespacedName. Even better:

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/discovery/kubernetes
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
                 │   olddisc    │              newdiscv2              │
                 │    sec/op    │   sec/op     vs base                │
ResolvePodRef-16   516.3n ± 17%   265.3n ± 4%  -48.62% (p=0.000 n=10)

                 │   olddisc    │             newdiscv2              │
                 │     B/op     │    B/op     vs base                │
ResolvePodRef-16   1168.00 ± 0%   16.00 ± 0%  -98.63% (p=0.000 n=10)

                 │  olddisc   │             newdiscv2              │
                 │ allocs/op  │ allocs/op   vs base                │
ResolvePodRef-16   2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
```